refactor(user): moved repository folder
Moved repository folder to implementation folder inside infra layer
